backend/requirements.txt
fastapi==0.115.2
uvicorn==0.32.0
SQLAlchemy==2.0.36
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
itsdangerous==2.2.0
email-validator==2.2.0
